Language changes for #165.
Made some changes to the language files, some of which are to fix grammatical errors, some of which are to add (what I hope are helpful) explanations, some of which are to remove redundant or unnecessary wording, and some of which are simply (in my opinion) improved phrasings.

Probably the most significant part of these changes is that I have introduced the term "Trophy post" to describe the post with most likes/thanks for each user. If you guys don't like that term, we can change it.

Also made changes to README.md, the most significant of which is to expand on the upgrade procedure, explaining that when new settings are involved, the plugin must first be uninstalled prior to upgrade, not just deactivated.