Add hooks for #172: like post to show content.
Add two hooks to support external developers in implementing the
"thank/like post to show its content" feature.

The first hook is server-side, in the main thankyoulike.php file:
`thankyoulike_ajax_end`. It can be used to insert the contents of the
post into the returned JSON data.

The second hook is client-side, in jscripts/thankyoulike.min.js:
`tylAddDoneHooks`, an array. It can be used to insert and display the
contents of the liked post. An example of how to hook in is as
follows (this Javascript MUST be placed AFTER the <script> tag which
includes `/jscripts/thankyoulike.min.js` in the `showthread` template):

<script type="text/javascript">
if (typeof tylAddDoneHooks !== 'undefined')  {
	tylAddDoneHooks[tylAddDoneHooks.length] = function(data, pid) {
			console.log('Like added! The pid of the liked post is: '+pid);
			console.log('The returned data for the liked post is:');
			console.log(data);
	}
}
</script>

Additionally, add a client-side hook `tylDelDoneHooks`
simply for symmetry, in case it is ever useful/necessary.