Resolve #153: show new tyls since last click in alerts (#263)
When a new tyl is added to a post, update any tyl alert for that post
rather than creating a new alert. Show the total number of tyls for the
post as well as the total number of new tyls since the author last
clicked the tyl alert for the post.