Fully fix #295 by removing creation of new indexes from auto-upgrade (#300)
PR #296, which supports semi-automated creation of the indexes
via the plugin's ACP listing - was not enough. The full fix is
to totally remove the auto-creation of the indexes on plugin
auto-upgrade, and rely on users semi-automatically creating
them per the feature introduced by that PR.

This is because if/when the index creation times out on plugin
auto-upgrade, the upgrade process will not have completed, and
the admin might not know to repeat it by deactivating and then
reactivating the plugin after manually creating the indexes.