Refix #348: Traces! (#350)
This refix:

1. Recognises that prior even to the first fix, the plugin already
hooked in to a core member deletion routine. That hook -
`admin_user_users_delete_commit` - though does not run in all scenarios
in which users are deleted; notably, it does not run in scenarios in
which multiple users are deleted at once. In this commit, then, we
delete that now redundant preexisting hook-in which is superseded by the
more generic one which we added in the first fix.

2. Replaces in that new, generic hook-in the call to our generic recount
function with ad-hoc code after all, because, hey, why not be more
efficient?

3. Handles the updating of unread alerts that are affected by having had
tyls deleted due to their tyl'ing member having been deleted.

Refixes #348